src/services/reviews/requirements.txt
python-dotenv
azure-cosmos
fastapi[all]
requests